refactor(chat): split default-layout chat composites
